+++ resources/js/components/catalog-new/SearchOptionsController.vue
